update ntfs and ksmbd

+++ fs/smb/server/smb2misc.c
@@ -6,6 +6,7 @@
 
 #include "glob.h"
 #include "nterr.h"
+#include "smb2pdu.h"
 #include "smb_common.h"
 #include "smbstatus.h"
 #include "mgmt/user_session.h"
@@ -178,7 +179,7 @@ static int smb2_get_data_area_len(unsigned int *off, unsigned int *len,
 		break;
 	}
 	case SMB2_IOCTL:
-		*off = max_t(unsigned int,
+		*off = max_t(unsigned short int,
 			     le32_to_cpu(((struct smb2_ioctl_req *)hdr)->InputOffset),
 			     offsetof(struct smb2_ioctl_req, Buffer));
 		*len = le32_to_cpu(((struct smb2_ioctl_req *)hdr)->InputCount);
